+++ src/components/Checkbox/index.js
@@ -18,7 +18,7 @@ const Checkbox = (props) => {
   });
   return (
     <Label testSection={ props.testSection && props.testSection + '-label' }>
-      <div className="flex flex-align--center">
+      <div className="flex">
         <input
           type="checkbox"
           defaultChecked={ props.defaultChecked }
@@ -26,6 +26,7 @@ const Checkbox = (props) => {
           disabled={ props.isDisabled }
           onChange={ props.onChange }
           data-test-section={ props.testSection }
+          style={ { 'marginTop': '0.35em' } }
         />
         <div className={ labelClassNames }>
           { props.label }