+++ src/components/Radio/index.js
@@ -18,7 +18,7 @@ const Radio = (props) => {
   });
   return (
     <Label testSection={ props.testSection && props.testSection + '-label' }>
-      <div className="flex flex-align--center">
+      <div className="flex">
         <input
           type="radio"
           name={ props.name }
@@ -27,6 +27,7 @@ const Radio = (props) => {
           disabled={ props.isDisabled }
           onChange={ props.onChange }
           data-test-section={ props.testSection }
+          style={ { 'marginTop': '0.3em' } }
         />
         <div className={ labelClassNames }>
           { props.label }